CRB Tech Reviews - The Inside Story
Owing to the ongoing COVID-19 pandemic, the education sector has seen a major transformation. It has switched to the online mode of training. CRBtech, one of the top and established edutech companies having an experience of over 15 years in the training and placement sector, is no exception to this. It has also adapted to this change by coming up with online training programs in different leading domains under CRBneo.